services/users/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose[cryptography]   
PyJWT                        
pydantic
python-dotenv
aioboto3                     
pillow                       
httpx                        
python-multipart             
